fix: "Update doc-test to ignore checks in CI environment
Modified doc-tests to ignore certain checks when tests are running in a CI environment. The CI environment sometimes can produce different results, so for consistency tests involving `os.environ.get()` and `parse_env("PATH")` are only performed if not in a CI environment.

Also fixed wrong module path in import statements and modified some tests to use ELLIPSIS flag to match different path lengths. Added the CI and ENV environment variables to the doctest_namespace fixture for tests to use."